requirements.txt
pydantic>=2.0.0
openai>=1.0.0
httpx>=0.25.0
networkx>=3.0
matplotlib>=3.7.0
pytest>=8.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0

